backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]